Refactor getDisplayedFilters and add translations
